+++ CHANGELOG.rst
@@ -7,6 +7,9 @@ Note that ``PHAB_ID=#`` and ``RB_ID=#`` correspond to associated messages in com
 Unreleased
 ----------
 
+19.8.0
+------
+
 * scrooge-generator: The deprecated `ReqRepThriftServiceBuilder` has been
   removed. ``PHAB_ID=D345740``